+++ packages/bricks/src/Badge.tsx
@@ -43,9 +43,9 @@ const Badge = forwardRef<"span", BadgeProps>((props, forwardedRef) => {
 	return (
 		<Role.span
 			{...rest}
-			data-kiwi-tone={tone}
-			data-kiwi-variant={variant}
-			className={cx("🥝-badge", props.className)}
+			data-_sk-tone={tone}
+			data-_sk-variant={variant}
+			className={cx("🥝Badge", props.className)}
 			ref={forwardedRef}
 		>
 			{label}