+++ packages/bricks/src/Progress.tsx
@@ -81,7 +81,7 @@ const ProgressBar = forwardRef<"div", ProgressBarProps>(
 			return value != null
 				? {
 						...styleProp,
-						"--🥝progress-bar-fill-portion-shown": `${value}%`,
+						"--🥝ProgressBar-fill-portion-shown": `${value}%`,
 					}
 				: styleProp;
 		}, [styleProp, value]);
@@ -93,10 +93,10 @@ const ProgressBar = forwardRef<"div", ProgressBarProps>(
 				aria-valuemin={value != null ? 0 : undefined}
 				aria-valuemax={value != null ? 100 : undefined}
 				{...rest}
-				data-kiwi-size={size}
-				data-kiwi-tone={tone}
-				data-kiwi-variant={value != null ? "determinate" : "indeterminate"}
-				className={cx("🥝-progress-bar", props.className)}
+				data-_sk-size={size}
+				data-_sk-tone={tone}
+				data-_sk-variant={value != null ? "determinate" : "indeterminate"}
+				className={cx("🥝ProgressBar", props.className)}
 				style={style}
 				ref={forwardedRef}
 			/>