+++ packages/bricks/src/Skeleton.tsx
@@ -42,9 +42,9 @@ const Skeleton = forwardRef<"div", SkeletonProps>((props, forwardedRef) => {
 		<Role.div
 			{...rest}
 			ref={forwardedRef}
-			className={cx("🥝-skeleton", props.className)}
-			data-kiwi-variant={variant}
-			data-kiwi-size={size}
+			className={cx("🥝Skeleton", props.className)}
+			data-_sk-variant={variant}
+			data-_sk-size={size}
 			aria-hidden
 		/>
 	);