+++ packages/foundations/src/Root.tsx
@@ -67,10 +67,15 @@ interface RootProps extends BaseProps {
 	 * ```
 	 */
 	unstable_htmlSanitizer?: (html: string) => string;
+
+	/**
+	 * Allows to customize the root portal container element.
+	 */
+	portalContainer?: React.ReactElement;
 }
 
 /**
- * Component to be used at the root of your application. It ensures that kiwi styles and fonts are loaded
+ * Component to be used at the root of your application. It ensures that StrataKit styles and fonts are loaded
  * and automatically applied to the current page or the encompassing shadow-root.
  *
  * Make sure to specify the `colorScheme` and `density` props.
@@ -89,6 +94,7 @@ export const Root = forwardRef<"div", RootProps>((props, forwardedRef) => {
 		children,
 		synchronizeColorScheme = false,
 		unstable_htmlSanitizer = identity,
+		portalContainer: portalContainerProp,
 		...rest
 	} = props;
 
@@ -109,6 +115,7 @@ export const Root = forwardRef<"div", RootProps>((props, forwardedRef) => {
 				colorScheme={props.colorScheme}
 				density={props.density}
 				ref={setPortalContainer}
+				render={portalContainerProp}
 			/>
 
 			<PortalContext.Provider value={portalContainer}>
@@ -146,9 +153,9 @@ const RootInternal = forwardRef<"div", RootInternalProps>(
 		return (
 			<Role
 				{...rest}
-				className={cx("🥝-root", props.className)}
-				data-kiwi-theme={colorScheme}
-				data-kiwi-density={density}
+				className={cx("🥝Root", props.className)}
+				data-_sk-theme={colorScheme}
+				data-_sk-density={density}
 				ref={useMergedRefs(forwardedRef, findRootNodeFromRef)}
 			>
 				<RootNodeContext.Provider value={rootNode}>
@@ -191,33 +198,47 @@ function SynchronizeColorScheme({
 
 // ----------------------------------------------------------------------------
 
+interface PortalContainerProps
+	extends Pick<RootProps, "colorScheme" | "density" | "render"> {}
+
 /** A separate root rendered at the end of root node, to be used as the container for all portals. */
-const PortalContainer = forwardRef<
-	"div",
-	Pick<RootProps, "colorScheme" | "density">
->((props, forwardedRef) => {
-	const rootNode = useRootNode();
-	if (!rootNode) return null;
+const PortalContainer = forwardRef<"div", PortalContainerProps>(
+	(props, forwardedRef) => {
+		const rootNode = useRootNode();
+		if (!rootNode) return null;
 
-	const destination = isDocument(rootNode) ? rootNode.body : rootNode;
-	if (!destination) return null;
+		const destination = isDocument(rootNode) ? rootNode.body : rootNode;
+		if (!destination) return null;
 
-	return ReactDOM.createPortal(
-		<div
-			className="🥝-root"
-			data-kiwi-theme={props.colorScheme}
-			data-kiwi-density={props.density}
-			style={{ display: "contents" }}
-			ref={forwardedRef}
-		/>,
-		destination,
-	);
-});
+		return ReactDOM.createPortal(
+			<Role
+				render={props.render}
+				className="🥝Root"
+				data-_sk-theme={props.colorScheme}
+				data-_sk-density={props.density}
+				style={{ display: "contents" }}
+				ref={forwardedRef}
+			/>,
+			destination,
+		);
+	},
+);
 
 // ----------------------------------------------------------------------------
 
 function Styles() {
 	const rootNode = useRootNode();
+
+	useLayoutEffect(
+		/** Adds `@layer reset` _before_ all other styles to ensure correct layer order.  */
+		function addResetLayer() {
+			if (!rootNode) return;
+			const styleElement = document.createElement("style");
+			((rootNode as Document).head || rootNode).prepend(styleElement);
+			styleElement.textContent = "@layer reset;";
+		},
+		[rootNode],
+	);
 
 	useLayoutEffect(() => {
 		if (!rootNode) return;