+++ packages/structures/src/index.ts
@@ -7,8 +7,10 @@
 export * as unstable_AccordionItem from "./AccordionItem.js";
 export { default as unstable_Banner } from "./Banner.js";
 export { default as Chip } from "./Chip.js";
+export * as unstable_Dialog from "./Dialog.js";
 export * as DropdownMenu from "./DropdownMenu.js";
 export * as unstable_ErrorRegion from "./ErrorRegion.js";
+export * as unstable_NavigationRail from "./NavigationRail.js";
 export * as Table from "./Table.js";
 export * as Tabs from "./Tabs.js";
 export * as unstable_Toolbar from "./Toolbar.js";