+++ packages/compat/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.0.7
+
+- Updated dependencies:
+  - @stratakit/structures@0.4.0
+  - @stratakit/foundations@0.3.0
+  - @stratakit/bricks@0.4.0
+
+## 0.0.6
+
+- [#729](https://github.com/iTwin/design-system/pull/729): Added `Select` component.
+
 ## 0.0.5
 
 - [#726](https://github.com/iTwin/design-system/pull/726): Added `Alert` component.